自機の弾で相手の弾を消す
