+++ Assets/Prefabs/Bullet.prefab
@@ -23,6 +23,7 @@ GameObject:
   - component: {fileID: 50737466026629774}
   - component: {fileID: 212805906843858966}
   - component: {fileID: 114572188124309412}
+  - component: {fileID: 114341563262137030}
   m_Layer: 0
   m_Name: Bullet
   m_TagString: Bullet
@@ -78,7 +79,7 @@ CircleCollider2D:
   m_Offset: {x: 0, y: 0}
   serializedVersion: 2
   m_Radius: 0.5
---- !u!114 &114572188124309412
+--- !u!114 &114341563262137030
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
@@ -86,11 +87,26 @@ MonoBehaviour:
   m_GameObject: {fileID: 1124153292108740}
   m_Enabled: 1
   m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5cecad11ef5604177b09bcac6d0243ef, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  damage: 30
+  range: 100
+  end_perticle: {fileID: 1583133410359590, guid: cc6e41e2952486045bb23f8319fb10bd,
+    type: 2}
+  hitpoint: 10
+--- !u!114 &114572188124309412
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1124153292108740}
+  m_Enabled: 0
+  m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 1ce954f3b6b17466793361353a464d94, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   damage: 0
-  enemybullet: 1
   range: 100
   end_perticle: {fileID: 1583133410359590, guid: cc6e41e2952486045bb23f8319fb10bd,
     type: 2}
