+++ Assets/Prefabs/BulletBomb.prefab
@@ -23,6 +23,7 @@ GameObject:
   - component: {fileID: 50341337285126002}
   - component: {fileID: 212531264363897388}
   - component: {fileID: 114568939888983434}
+  - component: {fileID: 114652897000976560}
   m_Layer: 0
   m_Name: BulletBomb
   m_TagString: Bullet
@@ -53,7 +54,7 @@ Rigidbody2D:
   m_BodyType: 0
   m_Simulated: 1
   m_UseFullKinematicContacts: 0
-  m_UseAutoMass: 0
+  m_UseAutoMass: 1
   m_Mass: 1
   m_LinearDrag: 0
   m_AngularDrag: 0.05
@@ -84,15 +85,29 @@ MonoBehaviour:
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1388754235893686}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 0e7187e62e8ae4593ba1cef2b2d41f22, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   damage: 100
-  enemybullet: 0
   range: 10
   end_perticle: {fileID: 0}
+  hitpoint: 100
+--- !u!114 &114652897000976560
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1388754235893686}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b00de5b31e6b2437383d6b6e874c9f3d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  damage: 30
+  range: 100
+  end_perticle: {fileID: 100006, guid: 8a4aceaf2957c044480166b429f4c7d0, type: 2}
   hitpoint: 100
 --- !u!212 &212531264363897388
 SpriteRenderer:
