+++ Assets/Scenes/Game.unity
@@ -582,6 +582,11 @@ Prefab:
       propertyPath: m_RootOrder
       value: 2
       objectReference: {fileID: 0}
+    - target: {fileID: 96040968988493230, guid: c6e2912881c274e69a685abb204b9ec2,
+        type: 2}
+      propertyPath: m_Autodestruct
+      value: 1
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: c6e2912881c274e69a685abb204b9ec2, type: 2}
   m_IsPrefabParent: 0
