+++ Assets/Scripts/BulletController.cs
@@ -6,14 +6,10 @@ using UnityEngine.UI;
 public class BulletController : MonoBehaviour {
 
 	public int damage = 30;
-	public bool enemybullet = true;
-	public int range = 10;
+	public int range = 100;
 	public GameObject end_perticle;
 	public int hitpoint = 10;
 	private int distance = 0;
-
-	public void Start() {
-	}
 
 	public void Update() {
 
@@ -25,17 +21,6 @@ public class BulletController : MonoBehaviour {
 
 		// hitpoint 削られたら消える
 		if (hitpoint <= 0) {
-			if (end_perticle != null) {
-				Instantiate (end_perticle, transform.position, Quaternion.identity);
-			}
-			Destroy (this.gameObject);
-		}
-	}
-
-	// 衝突
-	public void OnTriggerEnter2D(Collider2D other) {
-		if ((enemybullet && other.tag == "Player") || (!enemybullet && other.tag == "Enemy")) {
-			other.gameObject.SendMessage ("setDamage", damage);
 			if (end_perticle != null) {
 				Instantiate (end_perticle, transform.position, Quaternion.identity);
 			}
