+++ Assets/Scripts/ExpandPlayerBullet.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: b00de5b31e6b2437383d6b6e874c9f3d
+timeCreated: 1509809534
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
