+++ Assets/Scripts/PlayerBullet.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 347576872b1594995a69c420affc8696
+timeCreated: 1509809402
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
