+++ Assets/Scripts/Weapon.cs
@@ -8,19 +8,14 @@ public class Weapon : MonoBehaviour {
 	public int damage = 20;
 	public int range = 10;
 	public float bulletsize = 1;
-	private bool enemybullet = true;
 	public AudioSource audioSource;
 
 	public void Start () {
-		if (tag == "Player") {
-			enemybullet = false;
-		}
 		audioSource = gameObject.GetComponent<AudioSource> ();
 	}
 
 	public GameObject shoot() {
 		GameObject obj = Instantiate (bullet, transform.position, Quaternion.identity);
-		obj.GetComponent<BulletController> ().enemybullet = enemybullet;
 		obj.GetComponent<BulletController> ().damage = damage;
 		obj.GetComponent<BulletController> ().range = range;
 		obj.GetComponent<BulletController> ().end_perticle = end_perticle;
